requirements.txt
bs4
notion-client

openai-whisper

langgraph

langchain_core
langchain
langchain_community
langchain_huggingface
langchain_openai
langchain_mistralai
langchain_gigachat

duckduckgo-search

faiss-cpu
pymupdf
Pillow
torch
transformers
protobuf 
sentencepiece
accelerate
grpcio
yandexcloud

telebot
telegramify-markdown
telegramify-markdown[mermaid]


pandas
